refactor: Organize code generator into modular package structure
Refactored tools/generate_models.py into a well-organized Python package
with modular components for better maintainability and reusability.

Changes:
- Split generate_models.py into focused modules:
  - _common.py: Base utilities with no dependencies
  - generators.py: Code generation functions
  - parsers.py: JSON/YAML parsing functions
  - type_utils.py: Type-related utilities
  - utils.py: Directory structure utilities
  - main.py: CLI entry point
  - __main__.py: Module entry point

- Updated documentation (README.md, AGENTS.md)
- Added package README with full documentation
- Updated .gitignore to ignore data directory

Closes #33
