+++ .gitignore
@@ -168,3 +168,4 @@ Temporary Items
 # Generated models
 src/armodel/models/generated/
 output.arxml
+data
