+++ tools/__init__.py
@@ -1,4 +1,17 @@
 """Code generation tools."""
-from .generate_models import parse_mapping_json, generate_class_code, generate_builder_code, generate_all_models, to_snake_case
 
-__all__ = ["parse_mapping_json", "generate_class_code", "generate_builder_code", "generate_all_models", "to_snake_case"]
+from .generate_models import (
+    generate_all_models,
+    generate_builder_code,
+    generate_class_code,
+    parse_mapping_json,
+)
+from .generate_models.utils import to_snake_case
+
+__all__ = [
+    "generate_all_models",
+    "generate_builder_code",
+    "generate_class_code",
+    "parse_mapping_json",
+    "to_snake_case",
+]
