+++ tools/generate_models/__main__.py
@@ -1,0 +1,6 @@
+"""Allow running the package as a module with python3 -m generate_models."""
+
+from generate_models.main import main
+
+if __name__ == "__main__":
+    main()
